Store a Value Into an RDA file

This stores or updates a value, its description, and its context within an RDA notebook file. Note that this will overwrite the results of any previous calls that defined this item.

storeRDA(
  name = NULL,
  value = NULL,
  comment = "",
  context = "",
  rda = whichRDA()
)

Arguments

name

character value indicating the name of the quantity to be stored. Note that this does not need to be a valid R identifier (see Example).

value

an R object storing the value to be stored.

comment

character value holding a comment about the quantity being stored.

context

character value indicating something about the context of the stored item. If this is an empty string (which is the default) then it is the current working directory as discovered with getwd(). If it is a string starting with the / character, then the current working directory is prepended to the string. Otherwise, `context' is used directly.

rda

character value indicating the name of the RDA file in which to save the notes. The default is the value returned by whichRDA(), as set by a previous call to useRDA().

Author

Dan Kelley

Examples

library(rdan)
useRDA("sample.rda")
storeRDA("2pi", 2 * pi, "circle perimeter divided by radius")
retrieveRDA("2pi")
#> $value
#> [1] 6.283185
#> 
#> $comment
#> [1] "circle perimeter divided by radius"
#> 
#> $context
#> [1] "/Users/kelley/git/rdan/docs/reference"
#> 
storeRDA("beatles on Ed Sullivan Show", as.Date("1964-02-09"), "Beatlemania starts")
retrieveRDA("beatles on Ed Sullivan Show")
#> $value
#> [1] "1964-02-09"
#> 
#> $comment
#> [1] "Beatlemania starts"
#> 
#> $context
#> [1] "/Users/kelley/git/rdan/docs/reference"
#> 
file.remove("sample.rda") # we don't need to store this!
#> [1] TRUE